+++ front-page-april-start.php
@@ -144,58 +144,6 @@
                 </div>
 
                 <div class="row mt-5 ml-0 mr-0 row-eq-height">
-
-                    <div class="col-md-6">
-
-                        <div class="block-picture-primary-container">
-
-                            <div class="block-container">
-                                <a href="<?php echo site_url('/brands/elk-lighting'); ?>">
-                                    <div class="block-picture">
-                                        <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="/wp-content/uploads/2020/05/89047_8.jpg">', wp_get_attachment_metadata(166726), 166726 ); ?>
-                                    </div>
-                                    <div class="block-caption-wrapper">
-                                        <div class="block-caption-container">
-                                            <h4 class="block-heading">Elk Lighting</h4>
-                                            <h3 class="block-heading">Get 15% off on All Elk Lighting</h3>
-                                            <h4 class="block-heading">Use coupon "ELKLIGHTING15" at checkout!</h4>
-                                            <a style="text-align: center;" href="<?php echo site_url('/brands/elk-lighting'); ?>" class="block-shop-link">Shop Now</a>
-                                        </div>
-                                    </div>
-                                </a>
-                            </div>
-
-                        </div>
-
-                    </div>
-
-                    <div class="col-md-6">
-
-                        <div class="block-picture-primary-container">
-
-                            <div class="block-container">
-                                <a href="<?php echo site_url('/brands/elk-home'); ?>">
-                                    <div class="block-picture">
-                                        <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="/wp-content/uploads/2020/07/elk_home_sale_august.jpg">', wp_get_attachment_metadata(172838), 172838 ); ?>
-                                    </div>
-                                    <div class="block-caption-wrapper">
-                                        <div class="block-caption-container">
-                                            <h4 class="block-heading">Elk Home</h4>
-                                            <h3 class="block-heading">Save 15% on all Elk Home Products</h3>
-                                            <h4 class="block-heading">Use coupon "ELKHOME15" at checkout!</h4>
-                                            <a style="text-align: center;" href="<?php echo site_url('/brands/elk-home'); ?>" class="block-shop-link">Shop Now</a>
-                                        </div>
-                                    </div>
-                                </a>
-                            </div>
-
-                        </div>
-
-                    </div>
-
-                </div>
-
-                <div class="row mt-5 ml-0 mr-0 row-eq-height">
                      
                     <div class="col-md-12 mb-3 mt-5">
                         <h2>Featured Designer Collections</h2>